rename universal methods file to avoid confusion with form_factor.universal
