+++ sparrowpy/form_factor/universal_methods.py
