+++ sparrowpy/radiosity_fast/source_energy.py
@@ -7,7 +7,7 @@ except ImportError:
     prange = range
 
 import numpy as np
-from sparrowpy.form_factor.universal import (
+from sparrowpy.form_factor.universal_methods import (
     pt_solution as patch2point)
